% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_key.R
\name{generate_key}
\alias{generate_key}
\title{Generate strain collection key}
\usage{
generate_key(path_experiment, path_control, collection_id)
}
\arguments{
\item{path_experiment}{String. Path to the experimental 96 well plate
CSV. The CSV should be in wide format with the first column labeled "row"
with corresponding row NUMBERS, and remaining columns numbered.}

\item{path_control}{String. Path to the control plate in positions 2 and 3.}

\item{collection_id}{String. Name of strain collection. Start with letter
and no spaces please!}
}
\description{
Generates strain collection keyfile assuming you are using
one 96 well plate, that is in position 1 and rotated 180 degrees
in position 4. Positions 2 and 3 are controls.
}
\details{
Woah, so many cool details about the implementation of this function.
Any special cases or things we should know?
}
\examples{
generate_key('Group_2.csv', 'WT.csv', 'MATa-CTF4-Group1')

}

